+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- cleaned up `MatrixRegex`
 - upgrade gradle to 8.9
 - close Okio streams by default
 - Make `refresh_token` required
@@ -26,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- parsing and scanning of mentions
 - fixed wrong calculation of canLoadBefore and canLoadAfter in Timeline, when room with upgrades is used
 - close JavaScript streams correctly